tree-wide: pass -std=gnu11 -Wvla
In LXC I've introduced the policy that the minimum supported compiler
version will be the same as for the kernel. This is currently gcc 4.8.
As such we can support the niceties that the C11 standard provides and
also disable the crazyness that are VLAs that have been marked optional
from C11 onwards. So always pass -std=gnu11 and -Wvla.

Closes #5066.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>