candid: Improve domain validation and pubkey
This introduces a new "candid.api.key" option which can be set to the
expected public key for the candid server.

Setting this key is effectively required if using a non-HTTPs server.
When using HTTPs, this is optional as LXD can fetch the key directly
from the candid server.

An issue with validation of domains is also fixed, effectively
preventing a candid token which doesn't include a domain for being
considered valid when a list of valid domains is set.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>