global: Implement LXD_INSECURE_TLS env variable
This introduces a new `LXD_INSECURE_TLS` environment variable which when
set to true will disable our restrictive list of ciphers for client to
server communications as well as image and instance type retrieval.

This effectively makes it possible to perform downgrade attacks against
LXD, breaking perfect forward secrecy and isn't something that should
ever be used unless your environment gives you no choice
(intercepting TLS proxy).

Communications between LXD servers (container or storage migration) as
well as communications within a LXD cluster WILL NOT respect this
environment variable and will keep enforcing a very strict list of
ciphers. That's because those communications should never be proxied in
the first place.

This is only available through environment variables due to how widely
used and deep our TLS code is and the need for both the CLI and daemon
to know about this (so a daemon config option wouldn't be sufficient).

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>