Prevent event listeners from lying around even after Disconnect()
Currently the EventListener.Disconnect() method just removes the listener from
the list, but does not close the underlying websocket connection (since there
might be other listeners). However if the list of event listeners has become
empty, the current logic won't close the connection since it's blocked on
conn.ReadMessage().

This change separates the job of fetching events via conn.ReadMessage() from the
job of monitoring if all listeners are gone.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>