shared: Read certificates from host
This is for snap users, during update of the core snap, the custom mount
namespace setup by LXD may be missing, causing access to the host CA
configuration to be temporarily broken (until the next LXD restart).

With this patch, we directly go look for the CA certificates of the host
and merge those with what's in the snap environment.

Closes https://github.com/lxc/lxd-pkg-snap/issues/28

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>