lxd/cluster: Fix schema upgrades
While working on LXD 3.0 deb to LXD 3.8 snap upgrades, I ran into an
issue where loading the prepared statements was erroring out, causing
the daemon to exit.

That happens when the first node upgrades as the rest of the cluster is
still on the older schema version, causing the schema upgrade code to be
skipped. The prepared statements were then loaded regardless, causing
the crash when preparing them.

Instead, as soon as we know that the DB hasn't been upgraded, return and
let the retry logic happen. Once the remaining nodes have upgraded, the
schema update will happen and the prepared statements will be
re-processed.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>