images: Tar and compress in a combined stream when packing an image
By doing tar and compress in a combined stream, we require
1 less copy of the tarred image, and save time for IO for
writing and reading that file.

In my tests with an image of about 1.4GB when compressed this
saves 2.6GB of space used by the tar file, and reduces total
export time from approx 4m40s to 4m10s (30s quicker).

Signed-off-by: Joel Hockey <joelhockey@chromium.org>