lxd/backups: Attempt to delete storage on failure
This is rather ugly as there's no container struct around to pass to the
normal delete functions, but this will hopefully do the trick for now.

Once we get the new internal storage API, we should be able to change
this to returning a lower level storage driver and can then call its
delete function rather than ContainerDelete.

Closes #5597

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>