container/lxc: Records hotplugged p2p/bridged nic's host_name into volatile data
This allows hotplugged p2p/bridged nics to be able to have routes added to them without needing to restart or needing newer kernel features.

Removes population of m["host_name"] in fillNetworkDevice from volatile.%s.host_name for p2p/bridged nics, as that this was used during LXC setup and so what was potentially temporary volatile data during hotplug became permanent boot time data on restart.

This was not needed, as if the user has specified a config host_name then this will be used instead of volatile data.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>