networks: DNS clustered mode is correctly detected during LXD init
Ensures that DNS clustered mode is correctly enabled when starting the networks, even if the network is being started as part of LXD init before a cluster has been setup.

This has the effect of starting dnsmasq and forkdns with the correct settings for clustered mode.

It uses the node.ClusterAddress() to indicate an intention for the node to join a cluster at some point in the future.

Fixes #5805

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>