networks: Starts dnsmasq with --no-ping option to avoid delaying lease file writes
This can cause delays to writing the lease file when the container is started and stopped quickly, meaning it doesn't respond to icmp ping and dnsmasq delays updating the lease file for several seconds.

This in turn can cause issues with LXD parsing the lease file when a container is stopped and then deleted in order to release DHCP leases.

The result is that DHCP packets are queued up at dnsmasq and are processed after the container has been deleted, leaving orphaned leases.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>