seccomp: cause a default message to be sent
If liblxc is sending us messages of an indeterminate size we have two
options:
1. trigger a default response and _close_ the connection
2. trigger a default connection and _keep_ the connection

Option 1. means that we keep hanging the container for 30s on each
request. If liblxc is sending us garbage or a message of unknown size.

Options 2. means we keep handling the requests and responding with a
dummy response without doing anything. This will cause liblxc to respond
with the defaule ENOSYS response to the kernel. The container will not
be stalled.

I think option 1. is better than option 2.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>