seccomp: handle setxattr syscall
Sigh, until we have a method to:
- tell seccomp to resume a syscall if we tell it to do so
- have seccomp filter pointer arguments as well
we are stuck in the unfortunate situation that we can't nicely handler
syscalls with pointer arguments.
This brings us to setxattr() which is just such a syscall - YAY ME.

I want to be clear that what we do here is a best-effort to handle this
but it is not necessarily to be considered safe and this feature needs
to come with the appropriate big fat warning. I have no desire to get
CVEed because of this feature.

What I do is to add the following filter:
setxattr(const char *path /* we can't filter on this */,
         const char *name /* we can't filter on this */,
         const void *value /* we can't filter on this */,
	 size_t size /* we can filter on this since
		      * trusted.overlay.opaque needs to have a value of
		      * length 1 */,
         int flags /* we could filter on this but it's useless for us */)

which is in lxc filter terms:

notify setxattr [3,1,SCMP_CMP_EQ]

This should have us match very few setxattr syscalls hopefully.

For those setxattr that we catch that are not trusted.overlay.opaque but
still have a data size of 1, we __try__ (Please please, note the
__try__.) as best as we can to assume the credentials of the syscalling
task. This right now includes ids and all namespaces.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>