lxd/task: Attempt to run tasks on schedule
How long a task took to run shouldn't affect when it will run next, that
is, unless it took longer than the interval, in which case it should
just run immediately.

This behavior is effectively causing our cluster heartbeats to be
delayed to up to every 20s (which is the offline threshold) when we
would expect them running every 10s.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>