forksyscall: remove same_fsinfo() logic completely
The same_fsinfo() logic was needed when we didn't attach to the mount
namespace of the calling task.
Now it is actually becoming a problem. Think of the scenario where you
attach a storage volume to a container at /var/lib/. Now we want to
verify that the taret directory in which the device is supposed to be
created is on the same filesystem on the host and in the container. But
we won't be able to open the target directory from the host since it
only exists in the container's mount namespace because we attached a
mount. That means this logic is broken. Wipe it. It should be secure
since I changed the logic to always attach to the mount namespace.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>