tests: Avoid event forwarding race condition
There is a tiny potential race right between a cluster node joins and
the event forwarding being functional. If we immediately get requested
to do cross-node activity (in this case importing an image and creating
a container from it), there is a chance that one or multiple events
won't make it, leading to the client tool hanging (waiting for an event).

This problem should disappear when we rework event distribution to move
away from the current broadcast approach, at which point we can have
individual LXD nodes refuse to serve requests until they are connected
to the event hub.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>