device/proxy: Implements default listen file mode of 0644
The documentation states that the default listen file mode is 0755 if not supplied.

However the mode argument was currently being passed to forkproxy as empty string if not supplied.

There was also no mention of 755 or umask in forkproxy itself, leading me to think the default as documented was not implemented.

After discussions with stgraber, we agreed a sensible default of 0644.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>