device/device/utils/unix: Device management function rework
Updates UnixDeviceCreate() to return a UnixDevice struct with all the info about the created device.
This is to avoid having to have duplicate code elsewhere to stat the device to find out about it's properties.

Also adds unixDeviceSetup() and unixDeviceSetupCharNum() helper functions that will call UnixDeviceSetup() and then use the returned information to configure the RunConfig struct to instruct LXD to attach the device to the instance.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>