shared/util/linux: Removes Major and Minor functions
These functions are provided by the unix package, and return the major and minor numbers as uint32.

As part of this, GetFileStat has been modified such that it will returned a major and minor number of 0 rather than -1 when the file being requested is not a char or block device.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>