shared/containerwriter/container/tar/writer: Updates use of GetFileStat
GetFileStat now uses the unix package to get file device major and minor number and returns them as uint32.

As such we can no longer use -1 to indicate whether or not the device has a major or minor number.

However in WriteFile the default value for the DevMajor and DevMinor value in the Header struct seem to be 0 which will be the same value as returned by GetFileStat when the file is not a char or block device.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>