device: Modifies New function to return device even if validation fails
This is so the caller can decide whether to use the device or not based on the scenario.

In cases where a new version of LXD restricts device validation such that a previously valid config is now invalid, it is useful for LXD to still be able to cleanly stop and remove the device even if validation of the old config fails.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>