container/lxc: Removes volatile device keys when device is actually removed
Or when a device is updated such that its device type changes.

It will also ensure that if the device type remains the same but the new device's config replaces an old volatile key that the old volatile key is removed.

Fixes issue where hwaddr property was left even if device type changed.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>