device/device/utils/infiniband: Adds IB MAC functions
Supports long (20 byte) and short (8 byte) variants for validating and setting MAC addresses on infiniband devices.

The short form variant is used when you want to specify the last 8 bytes of the MAC address manually without having to worry about what the first 12 bytes are (these indicate the queue and subnet of the device and can not be changed by the ip tool).

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>