device/config/devices: Changes Devices type to map[string]Device
In doing so, also adds functions to convert from native map[string]map[string]string type to Devices type and back to native.

Removes use of DeepCopy and dependency on shared package and replaces with the newly added Device.Clone() function.

Also moves ValidateDevice function to be attached to the Device type as the Validate() function for consistency.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>