shared/util: Removes ExtractSnapshotName
ExtractSnapshotName was dangerous to use as it did not validate the lenth of the slice returned when splitting the supplied volume name by the snapshot delimiter (/).

When used with untrusted inputs it could cause a crash.

Additionally it's functionality overlapped with the safer and more feature rich ContainerGetParentAndSnapshotName, so have removed in place of that.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>