lxd/storage: Fix custom volume with underscores
The recently added project logic in the volume code incorrectly assumes
that underscores are only valid as separate between project names and
volume names.

This isn't true for custom volumes which may already have undercores in
their names despite us not having support for custom volumes in projects.

When we do implement custom volumes in projects, we will rename all
volumes to include the project prefix at which point we will be able to
always split those and extract the project. Until then, only do project
splitting on containers.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>