lxd/backup/backup/instance/config: UpdateInstanceConfigStoragePool no longer updates snapshots backup.yaml
Several storage drivers do not make backup.yaml available during backup restore (for example BTRFS snapshots are read only, and ZFS needs to be mounted).

As such we should only update the pool name of the main volume's backup.yaml during backup restore.

We regenerate the backup.yaml file of snapshots when they are restored to the main instance instead.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>