lxd/migrate/container: Add support for 2-phase sync in migrationSink.Do()
This fixes a bug where the criu property was not being populated in the migration response header.

This field is leveraged in the source to indicate whether the instance is running and whether a 2-phase sync is needed so its important this is sent back even if criu not being used.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>