lxd/storage/backend/lxd: Removes expansion of pool's volume config into volume config in newVolume()
Instead stores pool config inside volume to allow read-time expansion of config.

This was because during validation we should only be validating (and potentially removing keys during migration validation) the volume's config.

We need to ensure the volume's config stays intact and is not merged with the pool's volume config, so post-validation it can be used to store in database.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>