lxd/device/nic/routed: Improves IPv6 forwarding and proxy_ndp sysctl detection
Routed mode requires forwarding enabled for "all" interfaces to enable general packet forwarding. This is already enforced in liblxc but for usability is now detected in LXD.

Additionally, routed mode also requires proxy_ndp enabled on "all" interfaces as this is needed to ensure the kernel responds to unicast neighbour solicitations. Otherwise it rejects the request and causes the remote node to have to expire the neighbour cache and perform a multicast NDP solicitation, which causes periodic spikes in latency every 15-20s.

Fixes #6668

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>