lxd/instance/drivers/driver/qemu: Switch to unsafe async I/O mode on ZFS pools backed by loop files
Detects if the storage root disk is running on a ZFS pool backed by a loop file and switches to use unsafe cached I/O when using old ZFS version to avoid kernel hangs.

Additionally if a raw disk is attached to a VM that is hosted on a ZFS filesystem we switch to using writeback cached I/O and disable O_DIRECT semantics as this is not supported.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>