Improve a first impression error message
The first thing a new user might try to do is create a new container named after a distro: "ubuntu-18.04".

This fails with an error that the name is not valid, but wasn't clear about the problem. 

This clarifies that the container name is supposed to be a valid *DNS* hostname. Since dots seem like a common problem in potential container names, those are called out specifically as a problem.