lxd/instance/drivers/load: Pass copy of device config to device.Validate
Ensures a copy of the instance's device config is passed to device.Validate to ensure devices cannot modify the source config map.

This way any internal modifications a device may make to its config map are not reflected outside of the device.

This mirrors the existing usage of device.New().

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>