Revert "lxd/instance/drivers/driver/qemu: Fix go routine leak and hanging lxc clients"
This reverts commit 961a5bf4cba1d5c071a632c19f532444854ab89c.

This is because although it fixes the hanging lxc clients and leaking go routines, it also introduces a regression when using `lxc exec` without a single command that is expected to output data directly to the console.

With the escape sequence being sent, this was intermittently wiping out the output of the command.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>