lxd/storage/backend/lxd: Refuse to create storage pool if dir exists on disk
If the directory `/var/lib/lxd/storage-pools/<pool>` exists on disk, but pool doesn't exist in DB, if you try and create a storage pool of the same name and it fails (likely because the existing dir is non-empty) then the entire storage pool directory will be removed.

Changes check to refuse to create storage pool if directory already exists on disk.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>