lxd/storage/drivers/driver/lvm/volumes: Updates VolumeSnapshots to use lvs for snapshot list
Rather than listing the directories under /var/lib/lxd/storage/<pool>/<volume type>s-snapshots for snapshot lists, switches to using the lvs command to output logical volumes that match the prefix used for volume snapshots.

This is so if the volume has snapshots on the LVM storage layer, but the snapshot directory has been removed, they are still returned in the list.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>