lxd/storage/drivers/utils: Separates block file rounding logic into own function
Called roundVolumeBlockFileSizeBytes, which is still used inside ensureVolumeBlockFile, however now it can be used for comparison with requested size for on-disk file.

Also changes ensureVolumeBlockFile signature to just accept a size parameter, rather than taking from Volume, this function can take a size that is not yet in the Volume's config.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>