seccomp: handle hugetlbfs mount syscall interception
hugetlbfs already allocates new instances on every mount.
We can deal with the filesystem through existing mount interception.

We just need to add automatic handling of uid and gid mount options when
detecting a hugetlbfs mount. We also need to have hugetlbfs bypass the shifted
option as we don't want to put shiftfs on top of it when we already set the
correct uid/gid.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>