production-setup: add net.core.bpf_jit_limit and kernel.keys.maxbytes
/* kernel.keys.maxbytes */
When all containers share the same id mapping they will share a keyring. But
since each container will get their own session key it will be appended to the
keyring which will thus grow quite large. Thus the limit needs to be bumped.

/* net.core.bpf_jit_limit */
This is a limit on the size of eBPF JIT allocations which is usually set to
PAGE_SIZE * 40000. When your kernel is compiled with CONFIG_BPF_JIT_ALWAYS_ON=y
then /proc/sys/net/core/bpf_jit_enable is set to 1 and can't be changed. On
such kernels the eBPF JIT compiler will treat failure to JIT compile a bpf
program such as a seccomp filter as fatal when it would continue on another
kernel. On such kernels the limit for eBPF jitted programs needs to be
increased siginficantly.

Cc: Tobias Schüring <tobias@raidboxes.de>
Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>