ExecReaderToChannel: Prevent endless loops
Well, there's a bug in ExecReaderToChannel() where we can end up in an endless
loop when the error we get from poll() is non-transient. Realistically, when
poll() returns -1 it means something is so wrong we can't recover from it and a
next call to poll() won't be any better. If we get POLLNVAL that's pretty
obvious that the fd has been closed and so this is guaranteed to end up in an
endless loop. If we get POLLERR we could maybe maybe argue that there's
something we can do but realistically, not really. So exit in all those cases.

(Note, that we've never hit this problem before with containers. It suddenly
 started appearing when the lxd-agent became a thing.)

Closes #6870.
Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>