shared/util/linux: Updates ExecReaderToChannel to accept a finisher chan as struct{}
This is because the action of closing such a channel suffices for indicating the channel is finished.

It also means that the sender doesn't need to initialise the channel as a buffered channel, as multiple reads are allowed on a closed channel.

The empty struct{} is used to indicate that the contents of the read message has no meaning, just the action of reading from it is the indicator.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>