lxd/instances: Better use userRequested on Update
This updates all calls to Update() to properly set userRequested to true
if the config change comes through a user request.

If it doesn't, then ignore validation issues as reverting will be just
as broken and there's nothing the user can do at that point anyway.

This also removes the logic checking if `volatile.` keys have been
modified since the check never actually applied to PUT/PATCH requests
due to us having to use that API when handling migrations.

So in practice userRequested was never set to true when a change was
actually requested by the user, making the check useless.

Closes #6661

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>