lxd/storage/drivers/zfs/volumes: Updates CreateVolume to allow regeneration of deleted image volumes
Detects if the deleted image volume about to be restored is larger than the current pool volume.size setting and if so, renames it to a random deleted name so that it is not used for restoration.

This allows a new volume to be created from the filler function, allowing the current pool volume.size setting to be applied to the newly created cached image volume.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>