lxd/storage/drivers: Fixes custom volume root mount perm issue for BTRFS and DIR
The mount-time EnsureMountPath() call was resetting the permission of the custom volume root path to 0711.

However for BTRFS and DIR drivers, because there is no actual mount occurring, this was resetting any permission set inside the instance.

For these drivers we now do not call EnsureMountPath if the path exists and the volumne type is custom.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>