Stop using Driver.SetContextTimeout() which is a no-op
The Driver.SetContextTimeout() has been a no-op for a while now and should be
dropped.

The functionality has been moved to the WithContextTimeout() option, which we
use when creating the driver object and that sets a 30 seconds timeout for
DB.Begin().

In all other cases we don't actually timeout anymore and rather rely on the
network layer to report a broken connection. So the original #5234 issue can't
happen anymore for non-clustered deployments.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>