lxd/operations/operations: Fix hanging cancelation
This prevents an operation, which has onCancel set, to block
indefinitely.

The issue was that the goroutine writes to chanCancel and then calls
op.done() which sets onCancel to nil. The main function checks for
onCancel == nil, and also tries to write to chanCancel. This call would
block as the size of the buffered channel is 1.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>