lxd/device/device/load: Removes devTypes map and updates load to use NICType function
Switches to an multi-level case statement inside `load` to simplify what was starting to become an overcomplicated device load process, due to the new requirement to perform a DB lookup when resolving NIC types.

This approach keeps all instantiation of different device types together in one place, and combined with the removal of the helper functions is equivalent amounts of code.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>