test: Wait for operations to be removed from the database
Pending operations take up to 5 seconds to be removed from the database, even
after completion. Since we don't record the status in the database, and that
status appears to be something that only the local node stores, we can't cheaply
know what the status is when we querying the database to decide placement of a
new container.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>