lxd/dnsmasq/dhcpalloc: Adds static DHCP allocation package for dnsmasq
This package allows one to allocate a static IP allocation (both IPv4 and IPV6) in dnsmasq's config for a particular hostname & mac combination.

This decouples the old allocation logic that was in device/nic_bridged.go from the IP filtering concepts for which it was originally added.

Now the plan is to use this same logic for OVN uplink port allocation on external LXD bridges.

This package provides a transaction style logic, to allow per-protocol allocation logic in the calling code, without incurring the overhead of having to parse the existing dnsmasq leases file once per protocol.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>