seccomp: enable bpf in unprivileged containers
Thing's aren't normal on Linux anymore and we have users that want to run bpf
in unprivileged containers. This is of course completely opt-in and not to be
used lightly.

This patch makes it possible for LXD to manage the bpf() syscalls of an
unprivileged task in a container.

For a start, this enables support for BPF_PROG_TYPE_CGROUP_DEVICE program types
and BPF_CGROUP_DEVICE attach and detach types. This serves as a POC that it is
possible to manage bpf programs for a task running in an unprivileged
container.

I expect this to be extended in the future should users have a need for it.
There are a several ways we could extend this:
- supporting more program types
- supporting maps of pre-compiled or pre-written bpf programs instead of
  loading the container's own
.
.
.

This requires an update to our production guide too, since bpf checks for
memlock limit and even root can't exceed it. So in order to make this work one
needs to set:

root	soft	memlock	unlimited	unset	maximum locked-in-memory address space (KB)
root	hard	memlock	unlimited	unset	maximum locked-in-memory address space (KB)

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>