forksyscall: use pidfds to attach to namespaces
This is currently somewhat penalizing kernels that don't have pidfd setns
support because we're trying to setns() via the pidfd first and then fallback
to regular setns(). It's also penalizing error paths even non newer kernels
because in that case we always try both because we can't discern based on
return value from setns() when passing a pidfd if it failed because it doesn't
supports pidfd or because an actual error happened.

So in the long-term we'll want a way for the LXD daemon to communicate the
feature mask it detected at startup time to its helpers. This could probably be
a 32 bitmask that records all the kernel features via flags and that is passed
as an argument to the subprocess which can check for features it needs to know
about. Or maybe something more clever.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>