lxd/network/driver/common: Moves notifier for delete into common
Ensures that notifications are sent *after* network has been deleted on local node where command was initiated,
this ensures that if there are singular shared activities that need to take place (such as removing OVN database config)
that influence the removal of the network on other nodes (i.e detecting if the parent uplink port is in use) that notifications
will be sent after these occur.

This fixes an issue where the OVN uplink ports on other cluster nodes were not correctly detected as "no in use" and removed because
the shared OVN network northbound database config had not yet been removed which meant an OVS patch port was still connected to the
uplink bridge, causing it to be considered "in use".

By removing the OVN NB database config first on the node where the network remove command was recieved, this will cause the OVS patch
ports to be removed on all nodes, so that when the notification arrives it will also cleanly remove any uplink OVS bridges and veth
ports that are not in use by other OVN networks.

This is also consistent with the update behavior, which also runs the local update tasks on the received node first, before notifying the other nodes.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>