seccomp: fix compilation on kernels without proper bpf.h
We use the BPF_DEVCG_DEV_CHAR macro as a cheap way to detect whether the kernel
has the correct headers available to be compiled for bpf support. Since cgo
doesn't have a good way of letting us probe for structs or enums the
alternative would be to vendor bpf.h similar to what we do for seccomp itself.
But that's annoying since bpf.h is quite large. So users that want bpf
interception support should make sure to have the relevant header available at
build time.

Closes: #7849
Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>