lxc: Better handle arguments
This allows things like:

 - lxc --project=blah shell abc
 - lxc --debug shell abc

Not that it aliases cannot support leading flags with value separated by
spaces as the alias handling logic does not know what flag is a boolean
and what flag takes a value.

As a result, use `lxc --project=blah shell` and not `lxc --project blah shell`
as the latter cannot work and will lead to a confusing error.

Closes #7625

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>