lxc/remote: Add project selection logic
This adds logic to configure a project at remote adding time.
The user can directly pass the wanted project with `--project NAME`.

If no name is passed and the default project is available, it will be used.
If the default project is not available and there is only one available
project, this will be used automatically.

If more than one project is available and the default project is not
available, then the user will be shown a list of projects and asked
which to use.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>