lxd/db/instances: Renames and reworks instanceListExpanded to InstanceList
Modifies it so that it accepts a callback function that can be used to process each instance.

This allows the callback function to perform any additional processing as needed.

Also moves it from ClusterTx to Cluster so that the call back functions can be run outside of the transaction in case they are slow.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>