lxd/cgroup: Implement proper typing
This changes the various functions to deal with int64 rather than
obscure strings in many cases. Also renames a few functions to have more
meaningful names and improves handling of -1/max for cgroup2

As part of this, memsw functions perform the needed math directly
instead of relying on the caller doing it.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>