lxd/network/driver/ovn: Updates InstanceDevicePortDelete to accept an instance UUID and a ovsExternalOVNPort hint
This allows OVN instance ports started using the old naming regime (using instance ID) to be properly cleaned up when they are stopped by utilising the current active OVN external switch port associated to the local OVS interface.

In cases where this is not available we fallback to using the latest regime of using the instance's UUID for generating the OVN switch port name.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>